requirements.txt
anthropic>=0.40.0
pyoverleaf>=0.1.0
pyyaml>=6.0
rich>=13.0.0
